# backend/openapi-schema.yml
openapi: 3.0.3
info:
  title: 高校选课系统 API
  version: 1.0.0
  description: 基于 Django REST Framework 的高校选课系统接口文档。
paths:
...
  /api/rounds:
    get:
      operationId: rounds_list
      description: 查询选课轮次。支持按学期筛选；传入 `paginate=1` 时，返回真实分页结构。
      summary: 获取轮次列表
      parameters:
      - in: query
        name: page
        schema:
          type: integer
        description: 页码，从 1 开始。仅在 `paginate=1` 时生效。
      - in: query
        name: page_size
        schema:
          type: integer
        description: 每页条数，默认 10，最大 100。仅在 `paginate=1` 时生效。
      - in: query
        name: paginate
        schema:
          type: string
        description: 传入 `1` 时启用分页返回结构。
      - in: query
        name: term_id
        schema:
          type: integer
        description: 按学期 ID 筛选。
      tags:
      - rounds
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoundListResponse'
          description: ''
    post:
      operationId: rounds_create
      description: 管理员创建选课轮次，并归属到指定学期。
      summary: 创建轮次
      tags:
      - rounds
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Round'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Round'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Round'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Round'
          description: ''
  /api/rounds/{round_id}:
    get:
      operationId: rounds_detail
      description: 根据轮次 ID 获取单个轮次详情。
      summary: 获取轮次详情
      parameters:
      - in: path
        name: round_id
        schema:
          type: integer
        required: true
      tags:
      - rounds
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Round'
          description: ''
    put:
      operationId: rounds_update
      description: 管理员更新指定轮次的时间、范围和限制条件。
      summary: 更新轮次
      parameters:
      - in: path
        name: round_id
        schema:
          type: integer
        required: true
      tags:
      - rounds
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Round'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Round'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Round'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Round'
          description: ''
...
  /api/terms:
    get:
      operationId: terms_list
      description: 查询学期列表。传入 `paginate=1` 时，返回真实分页结构。
      summary: 获取学期列表
      parameters:
      - in: query
        name: page
        schema:
          type: integer
        description: 页码，从 1 开始。仅在 `paginate=1` 时生效。
      - in: query
        name: page_size
        schema:
          type: integer
        description: 每页条数，默认 10，最大 100。仅在 `paginate=1` 时生效。
      - in: query
        name: paginate
        schema:
          type: string
        description: 传入 `1` 时启用分页返回结构。
      tags:
      - terms
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTermListResponse'
          description: ''
    post:
      operationId: terms_create
      description: 管理员创建新的学期。
      summary: 创建学期
      tags:
      - terms
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Term'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Term'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Term'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Term'
          description: ''
  /api/terms/{term_id}:
    get:
      operationId: terms_detail
      description: 根据学期 ID 获取单个学期详情。
      summary: 获取学期详情
      parameters:
      - in: path
        name: term_id
        schema:
          type: integer
        required: true
      tags:
      - terms
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Term'
          description: ''
    put:
      operationId: terms_update
      description: 管理员更新指定学期信息。
      summary: 更新学期
      parameters:
      - in: path
        name: term_id
        schema:
          type: integer
        required: true
      tags:
      - terms
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Term'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Term'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Term'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Term'
          description: ''
  /api/users:
    get:
      operationId: users_list
      description: 管理员查询系统用户。支持按角色和关键字筛选；传入 `paginate=1` 时，返回真实分页结构。
      summary: 获取用户列表
      parameters:
      - in: query
        name: page
        schema:
          type: integer
        description: 页码，从 1 开始。仅在 `paginate=1` 时生效。
      - in: query
        name: page_size
        schema:
          type: integer
        description: 每页条数，默认 10，最大 100。仅在 `paginate=1` 时生效。
      - in: query
        name: paginate
        schema:
          type: string
        description: 传入 `1` 时启用分页返回结构。
      - in: query
        name: q
        schema:
          type: string
        description: 关键字模糊搜索。
      - in: query
        name: role
        schema:
          type: string
        description: 按角色筛选，可选值：`student`、`teacher`、`admin`。
      tags:
      - users
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserListResponse'
          description: ''
    post:
      operationId: users_create
      description: 管理员创建学生、教师或管理员账号，并同步写入资料信息。
      summary: 创建用户
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminUserWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminUserWrite'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
  /api/users/{user_id}:
    get:
      operationId: users_detail
      description: 管理员根据用户 ID 查看单个用户详情。
      summary: 获取用户详情
      parameters:
      - in: path
        name: user_id
        schema:
          type: integer
        required: true
      tags:
      - users
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
    put:
      operationId: users_update
      description: 管理员更新指定用户的账号信息、资料字段和启用状态。
      summary: 更新用户
      parameters:
      - in: path
        name: user_id
        schema:
          type: integer
        required: true
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminUserWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminUserWrite'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
    delete:
      operationId: users_destroy
      description: 管理员删除指定用户，但不能删除当前登录账号。
      summary: 删除用户
      parameters:
      - in: path
        name: user_id
        schema:
          type: integer
        required: true
      tags:
      - users
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
components:
  schemas:
...
    AdminUserWrite:
      type: object
      properties:
        username:
          type: string
          title: 用户名
          description: 必填；长度为150个字符或以下；只能包含字母、数字、特殊字符“@”、“.”、“-”和“_”。
          pattern: ^[\w.@+-]+$
          maxLength: 150
        password:
          type: string
          writeOnly: true
          minLength: 6
        email:
          type: string
          format: email
          title: 电子邮件地址
          maxLength: 254
        role:
          $ref: '#/components/schemas/RoleEnum'
        real_name:
          type: string
          default: ''
        student_no:
          type: string
          default: ''
        teacher_no:
          type: string
          default: ''
        phone:
          type: string
          default: ''
        department:
          type: string
          default: ''
        is_enabled:
          type: boolean
          default: true
      required:
      - role
      - username
...
    PaginatedRoundListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Round'
        count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
      required:
      - count
      - page
      - page_size
      - results
      - total_pages
...
    PaginatedTermListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Term'
        count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
      required:
      - count
      - page
      - page_size
      - results
      - total_pages
    PaginatedUserListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/User'
        count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
      required:
      - count
      - page
      - page_size
      - results
      - total_pages
...
    RoleEnum:
      enum:
      - student
      - teacher
      - admin
      type: string
      description: |-
        * `student` - Student
        * `teacher` - Teacher
        * `admin` - Admin
    Round:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        term:
          type: integer
        term_name:
          type: string
          readOnly: true
        name:
          type: string
          maxLength: 100
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        target_scope:
          $ref: '#/components/schemas/TargetScopeEnum'
        max_courses:
          type: integer
          maximum: 4294967295
          minimum: 0
          format: int64
        allow_drop:
          type: boolean
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - end_at
      - id
      - name
      - start_at
      - term
      - term_name
      - updated_at
...
    TargetScopeEnum:
      enum:
      - all
      - student
      - teacher
      type: string
      description: |-
        * `all` - All
        * `student` - Student
        * `teacher` - Teacher
...
    Term:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - name
      - updated_at
...
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          title: 用户名
          description: 必填；长度为150个字符或以下；只能包含字母、数字、特殊字符“@”、“.”、“-”和“_”。
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          title: 电子邮件地址
          maxLength: 254
        role:
          type: string
          readOnly: true
        real_name:
          type: string
          readOnly: true
        student_no:
          type: string
          readOnly: true
        teacher_no:
          type: string
          readOnly: true
        phone:
          type: string
          readOnly: true
        department:
          type: string
          readOnly: true
        is_enabled:
          type: boolean
          readOnly: true
        is_staff:
          type: boolean
          title: 工作人员状态
          description: 指明用户是否可以登录到这个管理站点。
        is_superuser:
          type: boolean
          title: 超级用户状态
          description: 指明该用户缺省拥有所有权限。
        date_joined:
          type: string
          format: date-time
          title: 加入日期
      required:
      - department
      - id
      - is_enabled
      - phone
      - real_name
      - role
      - student_no
      - teacher_no
      - username
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT